+++ main.py
@@ -20,6 +20,11 @@ colours = {"wood": [120, 80, 7],"common": [207, 207, 207],
            "epic": [94, 11, 189], "legend": [212, 165, 23], 
            "mythic": [240, 67, 10],"exclusiveB": [46, 53, 71], 
            "exclusiveA": [46, 53, 71]}
+#strength for chopping wood
+axePower = {"none": [1, 5],"common": [2, 6], "uncommon": [4, 8], 
+            "rare": [10, 14], "epic": [14, 19], "legend": [18, 30], 
+            "mythic": [25, 50],"exclusiveB": [50, 100], 
+            "exclusiveA": [100, 200]}
 #player setup
 Player1 = player.Player(10, {"wood": 0, "common": 0, "uncommon": 0, 
                              "rare": 0, "epic": 0, "legend": 0, 
@@ -47,8 +52,10 @@ WoodII = quests.Quest("WoodII", "Collect 100 pieces of wood",
                       False, 100, 100, "wood")
 WoodIII = quests.Quest("WoodIII", "Collect 1000 pieces of wood", 
                        False, 1000, 1000, "wood")
+WoodIV = quests.Quest("WoodIV", "Collect 10000 pieces of wood", 
+                      False, 10000, 10000, "wood")
 #dict of quests and if they can be taken in the guild
-questDict = {WoodI: True, WoodII: True, WoodIII: True}
+questDict = {WoodI: True, WoodII: True, WoodIII: True, WoodIV: True}
 
 # Functions -------------------------------------------------------------------
 def shopMenu():
@@ -58,7 +65,7 @@ def shopMenu():
         print("-inventory\n-buy\n-exit")
         shopChoice = input("What do you want to do: ").lower()
         if shopChoice == "inventory":
-            Player1.printInv()
+            Player1.printInv(axePower)
         elif shopChoice == "buy":
             while True:
                 print(f"\nYou have {Player1.coins} coins")
@@ -105,7 +112,7 @@ def guildMenu():
               "-view accepted quests\n-turn in quests\n-exit")
         guildChoice = input("What do you do: ").lower()
         if guildChoice == "inventory":
-            Player1.printInv()
+            Player1.printInv(axePower)
         elif guildChoice == "accept quests":
             while True:
                 print("\nHere are all the available quests")
@@ -200,9 +207,10 @@ def forestMenu():
         print("-inventory\n-chop wood\n-exit")
         forestChoice = input("What do you do: ").lower()
         if forestChoice == "inventory":
-            Player1.printInv()
+            Player1.printInv(axePower)
         elif forestChoice == "chop wood":
-            woodAmount = random.randint(1,5)
+            woodAmount = random.randint(axePower[Player1.equipped][0], 
+                                        axePower[Player1.equipped][1])
             Player1.inventory["wood"]+=woodAmount
             print(f"\nYou have gained {woodAmount} wood")
             for quest in Player1.acceptedQuests:
@@ -246,7 +254,7 @@ def mainMenu():
         print("-inventory\n-move\n-quit")
         mainMenuChoice = input("What do you want to do: ").lower()
         if mainMenuChoice == "inventory":
-            Player1.printInv()
+            Player1.printInv(axePower)
         elif mainMenuChoice == "move":
             movementMenu()
         elif mainMenuChoice == "quit":